Only generate template and example data when needed

+++ scripts/set-up-repo.sh
@@ -20,25 +20,29 @@ if [ "$SHACL_URL" != "" -a "$AP_URL" != "" ]; then
   echo "Adding AP and SHACL links to README"
   $SCRIPTS_DIR/update-readme.sh $ap_url $shacl_url
 
-  cd $WORK_DIR
-  $SCRIPTS_DIR/clone-ap-data-to-dashboard.sh || exit 1
+  if [ -f "$ROOT_DIR/template.xlsx" ]; then
+    echo "Template and example data already exist"
+  else
+    cd $WORK_DIR
+    $SCRIPTS_DIR/clone-ap-data-to-dashboard.sh || exit 1
 
-  echo "Copying shacl file in-shacl directory"
-  cp $ROOT_DIR/in-shacl/* $WORK_DIR/ap-data-to-dashboard/in-shacl || exit 1
+    echo "Copying shacl file in-shacl directory"
+    cp $ROOT_DIR/in-shacl/* $WORK_DIR/ap-data-to-dashboard/in-shacl || exit 1
 
-  cd $WORK_DIR
-  echo "Installing ap-data-to-dashboard"
-  $SCRIPTS_DIR/install-ap-data-to-dashboard.sh &> /dev/null || exit 1
+    cd $WORK_DIR
+    echo "Installing ap-data-to-dashboard"
+    $SCRIPTS_DIR/install-ap-data-to-dashboard.sh &> /dev/null || exit 1
 
-  echo "Generating template and example data"
-  cd ap-data-to-dashboard
-  node src/shacl-to-template.js || exit 1
+    echo "Generating template and example data"
+    cd ap-data-to-dashboard
+    node src/shacl-to-template.js || exit 1
 
-  echo "Moving template and example data"
-  cp in-shacl/template.xlsx $ROOT_DIR/template.xlsx
-  cp in-shacl/template.schema.json $ROOT_DIR/in-shacl/template.schema.json
-  rm -rf $ROOT_DIR/example-data && mkdir $ROOT_DIR/example-data
-  cp in-shacl/dummy* $ROOT_DIR/example-data
+    echo "Moving template and example data"
+    cp in-shacl/template.xlsx $ROOT_DIR/template.xlsx
+    cp in-shacl/template.schema.json $ROOT_DIR/in-shacl/template.schema.json
+    rm -rf $ROOT_DIR/example-data && mkdir $ROOT_DIR/example-data
+    cp in-shacl/dummy* $ROOT_DIR/example-data
+  fi
 else
   echo "ERROR: AP url or SHACL url is not defined. Please update the file \"urls.env\"."
   exit 1
